requirements.txt
httpx>=0.27
PyYAML>=6.0
grpcio>=1.65
protobuf>=4.25
